requirements.txt
ascii-magic==2.7.2
pillow==10.4.0
pytest==8.4.1
requests==2.32.5
